requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bce-python-sdk==0.9.56
bcrypt==5.0.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
easyocr==1.7.2
ecdsa==0.19.1
fastapi==0.127.0
filelock==3.20.1
fsspec==2025.12.0
future==1.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.12.0
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
modelscope==1.33.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
openai==2.14.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
paddleocr==3.3.2
paddlex==3.3.12
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
postgrest==2.27.0
prettytable==3.17.0
propcache==0.4.1
psutil==7.1.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.2.5
pypdfium2==5.2.0
pyroaring==1.0.3
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
realtime==2.27.0
requests==2.32.5
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scikit-image==0.26.0
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.50.0
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.12.20
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
